agrof-main/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
tensorflow>=2.12.0
flask>=2.2.0
flask-cors>=3.0.0
python-dotenv>=1.0.0
opencv-python>=4.7.0
pillow>=9.0.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.2.0
pyyaml>=6.0.0
dataset-tools>=0.1.0
tqdm>=4.65.0
gunicorn>=20.1.0
matplotlib>=3.6.0
seaborn>=0.12.0
requests>=2.28.0
werkzeug>=2.2.0
psutil>=5.9.0
